Added working explosions, moving platforms and chaser

+++ GXPEngine/Platform.cs
@@ -5,10 +5,9 @@ using GXPEngine;
 
 public class Platform : EasyDraw
 {
-    public Platform(int pX, int pY, int pWidth, int pHeight) : base(pWidth, pHeight)
+    public Platform(float pX, float pY, int pWidth, int pHeight) : base(pWidth, pHeight)
     {
         Clear(Color.White);
-        //SetOrigin(pWidth / 2, 0);
         SetXY(pX, pY);
     }
 }
